feat: change volume API to percentage format (0-100)
- Change get_volume/set_volume from f32 (0.0-1.0) to u32 (0-100)
- Change get_input_volume/set_input_volume similarly
- Update DeviceState.volume field from f32 to u32
- Update get_volume_and_mute to return percentage
- Update example_usage.py to match new API

+++ src/audio/mod.rs
@@ -37,13 +37,13 @@ pub struct DeviceState {
     #[pyo3(get)]
     is_default: bool,
     #[pyo3(get)]
-    volume: f32,
+    volume: u32,  // 音量百分比 0-100
     #[pyo3(get)]
     is_muted: bool,
 }
 
 impl DeviceState {
-    pub fn new(state: String, is_default: bool, volume: f32, is_muted: bool) -> Self {
+    pub fn new(state: String, is_default: bool, volume: u32, is_muted: bool) -> Self {
         Self { state, is_default, volume, is_muted }
     }
 }
